Renamed PostFeed to RecentPostFeed, moved under swarm-social
